+++ frontend/tests/run.ts
@@ -25,6 +25,7 @@ import "./jobDetail.test.js";
 import "./testDetail.test.js";
 import "./detailReadability.test.js";
 import "./dashboardOverview.test.js";
+import "./attentionGroups.test.js";
 import "./pullRequests.test.js";
 import "./pullRequestEscalation.test.js";
 import "./responsiveLayout.test.js";